requirements.txt
# Core dependencies
mcp>=1.0.0
aiohttp>=3.8.0
Pillow>=9.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
isort>=5.0.0
mypy>=1.0.0
flake8>=5.0.0